decodeZMortonPosition -> encode, minor refactors
