+++ src/constants/StructSizes.ts
