+++ src/store/stage-data/process-stage-polygon-file/NLPropConversionDefs.ts
@@ -1,6 +1,6 @@
 import { NLTextureDef } from '@/types/NLAbstractions';
 import O from '@/constants/StructOffsets';
-import S from '@/hooks/StructSizes';
+import S from '@/constants/StructSizes';
 import {
   getTextureColorFormat,
   getTextureSize,
