+++ src/store/stage-data/process-stage-polygon-file/scanModel.ts
@@ -1,4 +1,4 @@
-import S from '@/hooks/StructSizes';
+import S from '@/constants/StructSizes';
 import {
   nlMeshConversions,
   nlModelConversions,
