--- src/utils/textures/TextureColorFormat.ts
@@ -1,6 +1,0 @@
-export type TextureColorFormat =
-  | 'RGB565'
-  | 'RGB555'
-  | 'ARGB4444'
-  | 'ARGB8888'
-  | 'ARGB1555';
