+++ src/utils/textures/parse/index.ts
@@ -1,6 +1,6 @@
 export { default as argb1555ToRgba8888 } from './argb1555ToRgba8888';
 export { default as argb4444ToRgba8888 } from './argb4444ToRgba8888';
-export { default as decodeZMortonPosition } from './decodeZMortonPosition';
+export { default as encodeZMortonPosition } from './encodeZMortonPosition';
 export { default as getTextureColorFormat } from './getTextureColorFormat';
 export { default as getTextureSize } from './getTextureSize';
 export { default as getTextureWrappingFlags } from './getTextureWrappingFlags';
